+++ src/Ui/Bundle/AdminBundle/Form/Type/Product/Option/UpdateOptionType.php
@@ -71,6 +71,10 @@ class UpdateOptionType extends AbstractUpdateType
             ->add('subjectedToValidation', CheckboxType::class, [
                 'required' => false,
             ])
+            ->add(
+                'canScanParticipant', CheckboxType::class, [
+                'required' => false,
+            ])
             ->add('translations', CollectionType::class, [
                 'entry_type' => TranslationsType::class,
                 'label'      => false,